+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tauri-plugin-aptabase"
-version = "0.1.3"
+version = "0.1.4"
 license = "MIT"
 description = "Tauri plugin for Aptabase Analytics"
 authors = [ "Guilherme Oenning" ]
@@ -18,7 +18,7 @@ thiserror = "1.0"
 reqwest = { version = "0.11", features = ["json"] }
 time = { version = "0.3", features = ["formatting"]}
 os_info = "3"
-sys-locale = "0.2"
+sys-locale = "0.3"
 uuid = "1"
 wry = "0"
 log = "0.4"
