+++ src/commands.rs
@@ -1,59 +1,14 @@
-use log::debug;
-use serde_json::{json, Value};
-use tauri::{
-  command,
-  AppHandle,  Runtime, Window, State,
-};
-use time::{
-    OffsetDateTime, format_description::well_known::Rfc3339,
-};
+use serde_json::Value;
+use tauri::{command, State};
 
 use crate::state::AptabaseState;
-use crate::error::Error;
-
-type Result<T> = std::result::Result<T, Error>;
+use crate::track_event::internal_track_event;
 
 #[command]
-pub async fn track_event<R: Runtime>(
+pub fn track_event(
+  state: State<'_, AptabaseState>,
   name: &str,
   props: Option<Value>,
-  _app: AppHandle<R>,
-  _window: Window<R>,
-  state: State<'_, AptabaseState>,
-) -> Result<()> {
-
-  if state.config.app_key.is_empty() {
-    return Ok(());
-  }
-
-  let body = json!({
-      "timestamp": OffsetDateTime::now_utc().format(&Rfc3339).unwrap(),
-      "sessionId": state.eval_session_id(),
-      "eventName": name,
-      "systemProps": {
-          "osName": state.sys_info.os_name,
-          "osVersion": state.sys_info.os_version,
-          "locale": state.sys_info.locale,
-          "engineName": state.sys_info.engine_name,
-          "engineVersion": state.sys_info.engine_version,
-          "appVersion": state.app_version,
-          "sdkVersion": concat!(env!("CARGO_PKG_NAME"), "@", env!("CARGO_PKG_VERSION"))
-      },
-      "props": props
-  });
-
-  let url = state.config.ingest_api_url.clone();
-
-  let response = state.http_client.post(url).json(&body).send().await;
-  match response {
-      Ok(response) => {
-          if !response.status().is_success() {
-            debug!("failed to track_event with status code {}", response.status());
-          }
-      }
-      Err(err) => {
-        debug!("failed to track_event: {}", err.to_string());
-      }
-  }
-  Ok(())
+) {
+  internal_track_event(state, name, props)
 }
